+++ src/main/scala/org/podval/metadata/HasName.scala
@@ -10,11 +10,9 @@ trait HasName(nameOverride: Option[String]):
 object HasName:
 
   trait Enum extends HasName:
-    self: HasName =>
     final override protected def defaultName: String = this.toString
 
   trait NonEnum extends HasName:
-    self: HasName =>
     final override protected def defaultName: String = XmlParser.className(this.getClass)
 
   def mapByName[K <: HasName, M](
@@ -22,23 +20,30 @@ object HasName:
     metadatas: Seq[M],
     hasName: (M, String) => Boolean
   ): Map[K, M] =
-    val result: Seq[(K, M)] = metadatas.map(metadata =>
+    val pairs: Seq[(K, M)] = metadatas.map(metadata =>
       find(
         keys,
         metadata,
         hasName
       ) -> metadata
     )
-    checkNoUnmatchedKeys(keys.toSet -- result.map(_._1).toSet)
-    result.toMap
+    checkNoDuplicateKeys(pairs.map(_._1))
+    val result: Map[K, M] = pairs.toMap
+    checkNoUnmatchedKeys(keys.toSet -- result.keySet)
+    result
 
   def bind[K, M](
     keys: Seq[K],
     metadatas: Seq[M],
     getKey: M => K
   ): Map[K, M] =
-    val result: Map[K, M] = metadatas.map(metadata => getKey(metadata) -> metadata).toMap
-    checkNoUnmatchedKeys(keys.toSet -- result.keySet)
+    val pairs: Seq[(K, M)] = metadatas.map(metadata => getKey(metadata) -> metadata)
+    checkNoDuplicateKeys(pairs.map(_._1))
+    val result: Map[K, M] = pairs.toMap
+    val keySet: Set[K] = keys.toSet
+    checkNoUnmatchedKeys(keySet -- result.keySet)
+    val extra: Set[K] = result.keySet -- keySet
+    require(extra.isEmpty, s"Extra keys: $extra")
     result
 
   def find[K <: HasName](
@@ -49,12 +54,6 @@ object HasName:
     metadata = names,
     hasName = (names: Names, name: String) => names.hasName(name)
   )
-
-  def findByNames[K <: HasName](keys: Seq[K], names: Names): K =
-    val result: Seq[K] = keys.filter(key => names.hasName(key.name))
-    require(result.nonEmpty, s"Unmatched metadata: $names")
-    require(result.length == 1, s"Metadata matched multiple keys: $names")
-    result.head
 
   private def find[K <: HasName, M](
     keys: Seq[K],
@@ -68,3 +67,7 @@ object HasName:
 
   private def checkNoUnmatchedKeys[K](unmatchedKeys: Set[K]): Unit =
     require(unmatchedKeys.isEmpty, s"Unmatched keys: $unmatchedKeys")
+
+  private def checkNoDuplicateKeys[K](keys: Seq[K]): Unit =
+    val duplicates: Set[K] = keys.groupBy(identity).filter((_, ts) => ts.length > 1).keySet
+    require(duplicates.isEmpty, s"Duplicate keys: $duplicates")